# docs/openapi.json
{"openapi": "3.1.0", "info": {"title": "QR Lifecycle Platform", "version": "1.0.0"}, "paths": {"/api/auth/login": {"post": {"tags": ["auth"], "summary": "Login", "operationId": "login_api_auth_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/customers/": {"get": {"tags": ["customers"], "summary": "List Customers", "operationId": "list_customers_api_customers__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/CustomerOut"}, "type": "array", "title": "Response List Customers Api Customers  Get"}}}}}, "security": [{"HTTPBearer": []}]}, "post": {"tags": ["customers"], "summary": "Create Customer", "operationId": "create_customer_api_customers__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CustomerCreate"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CustomerOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/customers/{customer_id}": {"get": {"tags": ["customers"], "summary": "Get Customer", "operationId": "get_customer_api_customers__customer_id__get", "parameters": [{"required": true, "schema": {"type": "integer", "title": "Customer Id"}, "name": "customer_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CustomerOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}, "put": {"tags": ["customers"], "summary": "Update Customer", "operationId": "update_customer_api_customers__customer_id__put", "parameters": [{"required": true, "schema": {"type": "integer", "title": "Customer Id"}, "name": "customer_id", "in": "path"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CustomerUpdate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CustomerOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}, "delete": {"tags": ["customers"], "summary": "Delete Customer", "operationId": "delete_customer_api_customers__customer_id__delete", "parameters": [{"required": true, "schema": {"type": "integer", "title": "Customer Id"}, "name": "customer_id", "in": "path"}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/products/": {"get": {"tags": ["products"], "summary": "List Products", "operationId": "list_products_api_products__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ProductOut"}, "type": "array", "title": "Response List Products Api Products  Get"}}}}}, "security": [{"HTTPBearer": []}]}, "post": {"tags": ["products"], "summary": "Create Product", "operationId": "create_product_api_products__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductCreate"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/products/{product_id}": {"get": {"tags": ["products"], "summary": "Get Product", "operationId": "get_product_api_products__product_id__get", "parameters": [{"required": true, "schema": {"type": "integer", "title": "Product Id"}, "name": "product_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}, "put": {"tags": ["products"], "summary": "Update Product", "operationId": "update_product_api_products__product_id__put", "parameters": [{"required": true, "schema": {"type": "integer", "title": "Product Id"}, "name": "product_id", "in": "path"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductUpdate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}, "delete": {"tags": ["products"], "summary": "Delete Product", "operationId": "delete_product_api_products__product_id__delete", "parameters": [{"required": true, "schema": {"type": "integer", "title": "Product Id"}, "name": "product_id", "in": "path"}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/qrcodes/generate": {"post": {"tags": ["qrcodes"], "summary": "Generate Qrcode", "operationId": "generate_qrcode_api_qrcodes_generate_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_generate_qrcode_api_qrcodes_generate_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/QrOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/qrcodes/": {"get": {"tags": ["qrcodes"], "summary": "List Qrcodes", "operationId": "list_qrcodes_api_qrcodes__get", "parameters": [{"required": false, "schema": {"type": "string", "title": "Query"}, "name": "query", "in": "query"}, {"required": false, "schema": {"type": "boolean", "title": "Active"}, "name": "active", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/QrOut"}, "type": "array", "title": "Response List Qrcodes Api Qrcodes  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/qrcodes/{code_id}": {"get": {"tags": ["qrcodes"], "summary": "Get Qrcode", "operationId": "get_qrcode_api_qrcodes__code_id__get", "parameters": [{"required": true, "schema": {"type": "string", "format": "uuid", "title": "Code Id"}, "name": "code_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/QrOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}, "patch": {"tags": ["qrcodes"], "summary": "Update Qrcode", "operationId": "update_qrcode_api_qrcodes__code_id__patch", "parameters": [{"required": true, "schema": {"type": "string", "format": "uuid", "title": "Code Id"}, "name": "code_id", "in": "path"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/QrUpdateRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/QrOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/qrcodes/{code_id}/reuse/start": {"post": {"tags": ["qrcodes"], "summary": "Start Reuse", "operationId": "start_reuse_api_qrcodes__code_id__reuse_start_post", "parameters": [{"required": true, "schema": {"type": "string", "format": "uuid", "title": "Code Id"}, "name": "code_id", "in": "path"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ReuseStartRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/QrOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/qrcodes/{code_id}/reuse/history": {"get": {"tags": ["qrcodes"], "summary": "Reuse History", "operationId": "reuse_history_api_qrcodes__code_id__reuse_history_get", "parameters": [{"required": true, "schema": {"type": "string", "format": "uuid", "title": "Code Id"}, "name": "code_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ReuseHistoryOut"}, "type": "array", "title": "Response Reuse History Api Qrcodes  Code Id  Reuse History Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/qrcodes/test-decode": {"post": {"tags": ["qrcodes"], "summary": "Test Decode", "operationId": "test_decode_api_qrcodes_test_decode_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_test_decode_api_qrcodes_test_decode_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Response Test Decode Api Qrcodes Test Decode Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/scan": {"post": {"tags": ["scan"], "summary": "Register Scan", "operationId": "register_scan_api_scan_post", "parameters": [{"description": "M\u00e3 QR", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Code Id", "description": "M\u00e3 QR"}, "name": "code_id", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ScanCreateResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/scan/{code_id}": {"get": {"tags": ["scan"], "summary": "Scan Page", "operationId": "scan_page_api_scan__code_id__get", "parameters": [{"required": true, "schema": {"type": "string", "format": "uuid", "title": "Code Id"}, "name": "code_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"text/html": {"schema": {"type": "string"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/scans": {"get": {"tags": ["scan"], "summary": "List Scans", "operationId": "list_scans_api_scans_get", "parameters": [{"required": false, "schema": {"type": "string", "format": "uuid", "title": "Code Id"}, "name": "code_id", "in": "query"}, {"required": false, "schema": {"type": "string", "format": "date-time", "title": "Start"}, "name": "start", "in": "query"}, {"required": false, "schema": {"type": "string", "format": "date-time", "title": "End"}, "name": "end", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ScanEventOut"}, "type": "array", "title": "Response List Scans Api Scans Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/export/scans.csv": {"get": {"tags": ["scan"], "summary": "Export Scans", "operationId": "export_scans_api_export_scans_csv_get", "parameters": [{"required": false, "schema": {"type": "string", "format": "uuid", "title": "Code Id"}, "name": "code_id", "in": "query"}, {"required": false, "schema": {"type": "string", "format": "date-time", "title": "Start"}, "name": "start", "in": "query"}, {"required": false, "schema": {"type": "string", "format": "date-time", "title": "End"}, "name": "end", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/analytics/summary": {"get": {"tags": ["analytics"], "summary": "Analytics Summary", "operationId": "analytics_summary_api_analytics_summary_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AnalyticsSummaryResponse"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/analytics/qr/{code_id}": {"get": {"tags": ["analytics"], "summary": "Qr Analytics", "operationId": "qr_analytics_api_analytics_qr__code_id__get", "parameters": [{"required": true, "schema": {"type": "string", "format": "uuid", "title": "Code Id"}, "name": "code_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/QrAnalyticsResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/healthz": {"get": {"summary": "Healthcheck", "operationId": "healthcheck_healthz_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"AnalyticsSummaryResponse": {"properties": {"summary": {"$ref": "#/components/schemas/SummaryStats"}, "scans_by_day": {"items": {"$ref": "#/components/schemas/DailyScan"}, "type": "array", "title": "Scans By Day"}}, "type": "object", "required": ["summary", "scans_by_day"], "title": "AnalyticsSummaryResponse"}, "Body_generate_qrcode_api_qrcodes_generate_post": {"properties": {"data": {"type": "string", "title": "Data"}, "product_id": {"type": "integer", "title": "Product Id"}, "customer_id": {"type": "integer", "title": "Customer Id"}, "reuse_allowed": {"type": "boolean", "title": "Reuse Allowed", "default": false}, "options": {"type": "string", "title": "Options"}, "mask_image": {"type": "string", "format": "binary", "title": "Mask Image"}, "logo_image": {"type": "string", "format": "binary", "title": "Logo Image"}}, "type": "object", "required": ["data"], "title": "Body_generate_qrcode_api_qrcodes_generate_post"}, "Body_test_decode_api_qrcodes_test_decode_post": {"properties": {"image": {"type": "string", "format": "binary", "title": "Image"}}, "type": "object", "required": ["image"], "title": "Body_test_decode_api_qrcodes_test_decode_post"}, "CustomerCreate": {"properties": {"name": {"type": "string", "title": "Name"}, "email": {"type": "string", "format": "email", "title": "Email"}, "phone": {"type": "string", "title": "Phone"}, "meta": {"type": "object", "title": "Meta"}}, "type": "object", "required": ["name"], "title": "CustomerCreate"}, "CustomerOut": {"properties": {"name": {"type": "string", "title": "Name"}, "email": {"type": "string", "format": "email", "title": "Email"}, "phone": {"type": "string", "title": "Phone"}, "meta": {"type": "object", "title": "Meta"}, "id": {"type": "integer", "title": "Id"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["name", "id", "created_at"], "title": "CustomerOut"}, "CustomerUpdate": {"properties": {"name": {"type": "string", "title": "Name"}, "email": {"type": "string", "format": "email", "title": "Email"}, "phone": {"type": "string", "title": "Phone"}, "meta": {"type": "object", "title": "Meta"}}, "type": "object", "title": "CustomerUpdate"}, "DailyScan": {"properties": {"day": {"type": "string", "format": "date", "title": "Day"}, "count": {"type": "integer", "title": "Count"}}, "type": "object", "required": ["day", "count"], "title": "DailyScan"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "LoginRequest": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["email", "password"], "title": "LoginRequest"}, "ProductCreate": {"properties": {"sku": {"type": "string", "title": "Sku"}, "name": {"type": "string", "title": "Name"}, "description": {"type": "string", "title": "Description"}, "owner_customer_id": {"type": "integer", "title": "Owner Customer Id"}, "meta": {"type": "object", "title": "Meta"}}, "type": "object", "required": ["sku", "name"], "title": "ProductCreate"}, "ProductOut": {"properties": {"sku": {"type": "string", "title": "Sku"}, "name": {"type": "string", "title": "Name"}, "description": {"type": "string", "title": "Description"}, "owner_customer_id": {"type": "integer", "title": "Owner Customer Id"}, "meta": {"type": "object", "title": "Meta"}, "id": {"type": "integer", "title": "Id"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["sku", "name", "id", "created_at"], "title": "ProductOut"}, "ProductUpdate": {"properties": {"sku": {"type": "string", "title": "Sku"}, "name": {"type": "string", "title": "Name"}, "description": {"type": "string", "title": "Description"}, "owner_customer_id": {"type": "integer", "title": "Owner Customer Id"}, "meta": {"type": "object", "title": "Meta"}}, "type": "object", "title": "ProductUpdate"}, "QrAnalyticsResponse": {"properties": {"code_id": {"type": "string", "title": "Code Id"}, "product_name": {"type": "string", "title": "Product Name"}, "total_scans": {"type": "integer", "title": "Total Scans"}, "timeline": {"items": {"$ref": "#/components/schemas/QrTimelineEntry"}, "type": "array", "title": "Timeline"}}, "type": "object", "required": ["code_id", "total_scans", "timeline"], "title": "QrAnalyticsResponse"}, "QrOut": {"properties": {"code_id": {"type": "string", "format": "uuid", "title": "Code Id"}, "product_id": {"type": "integer", "title": "Product Id"}, "customer_id": {"type": "integer", "title": "Customer Id"}, "reuse_allowed": {"type": "boolean", "title": "Reuse Allowed"}, "reuse_cycle": {"type": "integer", "title": "Reuse Cycle"}, "active": {"type": "boolean", "title": "Active"}, "version": {"type": "integer", "title": "Version"}, "ecc": {"type": "string", "title": "Ecc"}, "image_url_png": {"type": "string", "title": "Image Url Png"}, "image_url_svg": {"type": "string", "title": "Image Url Svg"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}}, "type": "object", "required": ["code_id", "reuse_allowed", "reuse_cycle", "active", "version", "ecc", "image_url_png", "image_url_svg", "created_at", "updated_at"], "title": "QrOut"}, "QrTimelineEntry": {"properties": {"ts": {"type": "string", "format": "date-time", "title": "Ts"}, "event": {"type": "string", "title": "Event"}, "reuse_cycle": {"type": "integer", "title": "Reuse Cycle"}, "meta": {"type": "object", "title": "Meta"}}, "type": "object", "required": ["ts", "event", "reuse_cycle"], "title": "QrTimelineEntry"}, "QrUpdateRequest": {"properties": {"active": {"type": "boolean", "title": "Active"}, "product_id": {"type": "integer", "title": "Product Id"}, "customer_id": {"type": "integer", "title": "Customer Id"}, "reuse_allowed": {"type": "boolean", "title": "Reuse Allowed"}}, "type": "object", "title": "QrUpdateRequest"}, "ReuseHistoryOut": {"properties": {"cycle": {"type": "integer", "title": "Cycle"}, "reason": {"type": "string", "title": "Reason"}, "note": {"type": "string", "title": "Note"}, "ts": {"type": "string", "format": "date-time", "title": "Ts"}}, "type": "object", "required": ["cycle", "ts"], "title": "ReuseHistoryOut"}, "ReuseStartRequest": {"properties": {"reason": {"type": "string", "title": "Reason"}, "note": {"type": "string", "title": "Note"}}, "type": "object", "required": ["reason"], "title": "ReuseStartRequest"}, "ScanCreateResponse": {"properties": {"message": {"type": "string", "title": "Message"}, "product_name": {"type": "string", "title": "Product Name"}, "reuse_cycle": {"type": "integer", "title": "Reuse Cycle"}}, "type": "object", "required": ["message", "reuse_cycle"], "title": "ScanCreateResponse"}, "ScanEventOut": {"properties": {"id": {"type": "integer", "title": "Id"}, "code_id": {"type": "string", "format": "uuid", "title": "Code Id"}, "ts": {"type": "string", "format": "date-time", "title": "Ts"}, "ip": {"type": "string", "title": "Ip"}, "user_agent": {"type": "string", "title": "User Agent"}, "referer": {"type": "string", "title": "Referer"}, "approx_geo": {"type": "string", "title": "Approx Geo"}, "device": {"type": "string", "title": "Device"}, "reuse_cycle_at_scan": {"type": "integer", "title": "Reuse Cycle At Scan"}}, "type": "object", "required": ["id", "code_id", "ts", "reuse_cycle_at_scan"], "title": "ScanEventOut"}, "SummaryStats": {"properties": {"total_scans": {"type": "integer", "title": "Total Scans"}, "scans_today": {"type": "integer", "title": "Scans Today"}, "active_qr": {"type": "integer", "title": "Active Qr"}, "reuse_cycles": {"type": "integer", "title": "Reuse Cycles"}}, "type": "object", "required": ["total_scans", "scans_today", "active_qr", "reuse_cycles"], "title": "SummaryStats"}, "TokenResponse": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "title": "Token Type", "default": "bearer"}}, "type": "object", "required": ["access_token"], "title": "TokenResponse"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"HTTPBearer": {"type": "http", "scheme": "bearer"}}}}